build: ensure correct line endings
Because we develop on both Windows and *nix we've had issues with line endings being all over the place.
Especially with any files that are touched by PowerShell cmdlets (Set/Add-Content etc).
This results in builds that run on different OSes constantly reporting unexpected git diffs 😞
To combat this we forcefully set line endings on any (tracked) files that are created as part of the build.