Add Get-GitChanges cmdlet and fix line endings
* feat: Get-GitChanges cmdlet

This cmdlet will get the git status for a given repository and report back the changes in a consistent format.

* chore: spellings

* fix: use better formatting for git changes

* refactor: update classes

Move some of the classes around for GitStatus so we can get cleaner output

* build: Check for uncommitted changes

We now run the documentation generation as part of the build. We don't want there to be any uncommitted changes to the repo so check for them and fail the build if so.

* fix: update build name

* build: change execution order

* fix: don't resolve path as it might not exist!

* style: try to manage line endings...

* style: I think I got the syntax wrong...

* style: trying to fix line endings

* fix: line endings?

* style: line endings

* ensure eol=lf

* build: set fetch-depth.

This should hopefully stop us from picking up erroneous changes...

* build: try setting it to 0

* style: use crlf

As PowerShell generates the files they get CRLF by default. 😞

* build: update error message

* docs: update docs

* build: remove fetch-depth

* docs: update docs

* docs: line endings...

* style: ignore line endings to try and get builds passing

* style: fix line endings

* style: disable all line ending checks

* style: trying LF stuff again

* style: still trying to fix line endings

* style: fix line endings

* docs: update note

* style: attempting another fix

* style: try again

* style: use crlf

* try forcing crlf

* style: try to fix line endings

* style: line endings

* style: recreating file for EOL

* style: recreate with CRLF

* docs: try replacing docs with LF

* style: replace lf

* fix: line endings

* Don't create a newline at the end of the file

* test: add brand new docs for testing on.

* more test docs

* remove test files

* another document test

* update docs from windows

* remove test docs

* test setting LF on Windows

* more line endings

* more line ending testing

* more play with line endings

* docs: remove test documentation

* docs: move documentation to Docs directory

Most other projects out there seem to use "Docs" rather than ".docs" and it also results in some VSCode extensions picking up the docs better. 🎉

* style: force line endings to LF

* style: Set an editorconfig to assist with EOL's

We've had an absolute nightmare getting line endings sorted.
Due to a mixture of git and the way PowerShell currently handles line endings (it doesn't!)
Hopefully by setting an editorconfig we can ensure any manual edits are set to the correct endings.

* feat: Set-LineEndings cmdlet

New cmdlet to assist with setting the correct line endings.

* build: ensure correct line endings

Because we develop on both Windows and *nix we've had issues with line endings being all over the place.
Especially with any files that are touched by PowerShell cmdlets (Set/Add-Content etc).
This results in builds that run on different OSes constantly reporting unexpected git diffs 😞
To combat this we forcefully set line endings on any (tracked) files that are created as part of the build.

* docs: update docs

* docs: update documentation

* docs: remove weird slack link that appeared 🤔

* docs: update module help

* build: set line endings once