build: try GPG signing the stage-release commits
Because we generate the changelog/help page via GHA and we _require_ signed commits for merging PR's this helps us deal with that.
We don't want to give permission to circumvent BP rules to the bot user in case it gets compromised!
Having a compromised signing key feels like a safer loss 😂