feat: set standard location for user configs.
We've been meaning to do this for a while - but give a standard location that users can store their config files in and have the module auto-read them.
This way they don't have to specify the -ConfigFile parameters each time.
This hasn't been fully implemented in this module yet, but will be soon.